% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_button.R
\name{beam_downloadButton}
\alias{beam_downloadButton}
\title{Download button}
\usage{
beam_downloadButton(id, plot_out, device)
}
\arguments{
\item{id}{Id name of tabset panel.}

\item{plot_out}{lot to save.}

\item{device}{Parameter from \code{\link[ggplot2]{ggsave}}.}
}
\value{
\code{generate_downloadButton} returns the output of \code{downloadHandler}
for given plot and device.
}
\description{
Generates download button.
}
\details{
This function uses \code{\link[shiny]{downloadHandler}}.
}
