% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_help_files.R
\name{create_help_files_custom}
\alias{create_help_files_custom}
\title{Custom help files}
\usage{
create_help_files_custom(
  files,
  help_dir = "helpfiles",
  content = "This helpfile is not finished yet."
)
}
\arguments{
\item{files}{A character vector of names to use in creating help files.}

\item{help_dir}{A character string of the directory to use for help files.}

\item{content}{A character string to display in the automatically created help file.}
}
\description{
This function creates help files with custom content.
}
\details{
This function is based on \code{\link[shinyhelper]{create_help_files}}.
}
