% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{dt_format}
\alias{dt_format}
\title{Table widget}
\usage{
dt_format(dat, cols = colnames(dat))
}
\arguments{
\item{dat}{Dataset. A matrix or data frame.}

\item{cols}{Names of columns to display.}
}
\description{
\code{dt_format} is a function for creating a graphical widget containing a table with download buttons (Excel and CSV).
}
\details{
This function uses \code{\link[DT]{datatable}}.
}
