% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{extract_tt_data_row}
\alias{extract_tt_data_row}
\title{Tooltip data}
\usage{
extract_tt_data_row(
  hv,
  plot_info,
  plot_data,
  plot_type = "geom_point",
  tt_range = 5
)
}
\arguments{
\item{hv}{Hoover.}

\item{plot_info}{chosen mapping from the \code{data} object from the output of \code{\link[ggplot2]{ggplot_build}} for displayed plot.}

\item{plot_data}{data on which the ggplot is based}

\item{plot_type}{Type of plot corresponding to the data from first mapping.
Accepts either \code{'geom_point'}, \code{'geom_segment'}
or \code{'geom_col'}. Default \code{'geom_point'}.}

\item{tt_range}{A number denoting maximum distance between hoover and objects on the plot.
The row, whose distance to the hoover is less than \code{tt_range} will be selected from the whole data. Default 5.}
}
\value{
\code{prepare_tt_data} returns prepared data frame containing one record.
}
\description{
Prepares data in order that it may be displayed in tooltip.
The row from the data which fits to the hover the most will be selected.
}
\details{
This function filters one row of the imputed data in order that it
corresponds the most to the hoover coordinates.
}
