% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tooltip.R, R/generate_downloadButton.R
\name{generate_downloadButton}
\alias{generate_downloadButton}
\title{Download button}
\usage{
generate_downloadButton(id, plot_out, device)

generate_downloadButton(id, plot_out, device)
}
\arguments{
\item{id}{Id name of tabset panel.}

\item{plot_out}{Reactive. Plot to save.}

\item{device}{Parameter from \code{\link[ggplot2]{ggsave}}.}
}
\value{
\code{generate_downloadButton} returns the output of \code{downloadHandler}
for given plot and device.

\code{generate_downloadButton} returns the output of \code{downloadHandler}
for given plot and device.
}
\description{
Generates download button.

Generates download button.
}
\details{
This function uses \code{\link[shiny]{downloadHandler}}.

This function uses \code{\link[shiny]{downloadHandler}}.
}
