% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tooltip.R
\name{prepare_tt_data}
\alias{prepare_tt_data}
\title{Tooltip data}
\usage{
prepare_tt_data(hv, plot_out, plot_type = "geom_point")
}
\arguments{
\item{hv}{Hoover.}

\item{plot_out}{Reactive. Plot to extract data from.}

\item{plot_type}{Type of plot. Accepts either \code{'geom_point'}, \code{'geom_segment'}
or \code{'geom_col'}. Default \code{'geom_point'}.}
}
\value{
\code{prepare_tt_data} returns prepared data frame containing one record.
}
\description{
Prepares data in order that it may be displayed in tooltip.
}
\details{
This function filters one row of the imputed data in order that it
coresponds the most to the hoover coordinates.
}
