% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_SERVER.R
\name{tabsetPanel_SERVER}
\alias{tabsetPanel_SERVER}
\title{Server for tabset panel.}
\usage{
tabsetPanel_SERVER(
  id,
  plot_out,
  table_out,
  plot_type = "point",
  tt_content = NULL,
  tt_range = 5,
  helpfiles = "helpfiles"
)
}
\arguments{
\item{id}{Id of tabset panel.}

\item{plot_out}{Reactive. Plot to display in tab. The items from first mapping will be connected with tooltip.}

\item{table_out}{Reactive. Table to display in tab.}

\item{plot_type}{Type of plot corresponding to the data from first mapping. Accepts either \code{'geom_point'}, \code{'geom_segment'}
or \code{'geom_col'}. Default \code{'geom_point'}.}

\item{tt_content}{Optional.}

\item{tt_range}{A number denoting maximum distance between hoover and objects on the plot.
The row, whose distance to the hoover is less than \code{tt_range} will be selected from the whole data. Default 5.}

\item{helpfiles}{A character string denoting directory to save empty help files. Default \code{'helpfiles'}.}
}
\description{
Creates module server for tabset panel with plot and table.
}
