% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_panel.R
\name{beam_plot_panel_UI}
\alias{beam_plot_panel_UI}
\title{UI for tabset panel.}
\usage{
beam_plot_panel_UI(
  id,
  tab_plot = paste(id, "plot"),
  tab_table = paste(id, "data"),
  helpfiles = "helpfiles"
)
}
\arguments{
\item{id}{Id name of tabset panel.}

\item{tab_plot}{Character. Title of tab containing plot. Default to
\code{paste(id, "plot")}.}

\item{tab_table}{Character. Title of tab containing table. Default to
\code{paste(id, "data")}.}

\item{helpfiles}{A character string denoting directory to save empty help files. Default \code{'helpfiles'}.}
}
\description{
Creates module ui for tabset panel. It contains two tabs - with
plot and an elegant table with respective data.
}
\details{
\code{tabsetPanel_UI} provides three download buttons (png, svg and jpeg)
and tooltip for plot and helpers for both plot and table.
}
