% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tooltip.R
\name{beam_tooltip}
\alias{beam_tooltip}
\title{Generate tooltip}
\usage{
beam_tooltip(hv, plot_obj, plot_data, plot_type, tt_content, tt_range = 5)
}
\arguments{
\item{hv}{Hoover.}

\item{plot_obj}{ggplot object}

\item{plot_data}{data on which the ggplot is based}

\item{plot_type}{Type of plot. Accepts either \code{'geom_point'}, \code{'geom_segment'}
or \code{'geom_col'}. Default \code{'geom_point'}.}

\item{tt_content}{Content}

\item{tt_range}{A number denoting maximum distance between hoover and objects on the plot.
The row, whose distance to the hoover is less than \code{tt_range} will be selected from the whole data. Default 5.}
}
\description{
Generates tooltip .
}
