% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_panel.R
\name{dt_format}
\alias{dt_format}
\title{Table widget}
\usage{
dt_format(dat, colnames = colnames(dat))
}
\arguments{
\item{dat}{a data object - a matrix or a data frame.}

\item{colnames}{Names of columns to display in the table. By default it is
equal to the names of the columns from the \code{dat}. Parameter \code{colnames} is
correspond the argument \code{colnames} from \code{\link[DT]{datatable}}.}
}
\description{
\code{dt_format} is a function for creating a graphical
widget containing a table with download buttons (Excel and CSV).
}
\details{
This function uses \code{\link[DT]{datatable}}.
}
