% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_panel.R
\name{flip_ggplot_build}
\alias{flip_ggplot_build}
\title{Supplementary function for ggplot_build}
\usage{
flip_ggplot_build(ggplot_build_data)
}
\arguments{
\item{ggplot_build_data}{Element \code{data} from the output of \code{\link[ggplot2]{ggplot_build}}.}
}
\description{
Changes ggplot_build \code{data} output in order to make it
compatible with the plot when coordinates are flipped.
}
