% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{generate_tooltip}
\alias{generate_tooltip}
\title{Generate tooltip}
\usage{
generate_tooltip(hv, plot_out, plot_type, tt_content)
}
\arguments{
\item{hv}{Hoover.}

\item{plot_out}{Reactive. Plot to extract data from.}

\item{plot_type}{Type of plot. Accepts either \code{'point'}, \code{'comparison'}
or \code{'bar'}. Default \code{'point'}.}

\item{tt_content}{Content}
}
\description{
Generates tooltip .
}
