% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{numericInput_h}
\alias{numericInput_h}
\title{Numeric input with helper.}
\usage{
numericInput_h(inputId, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{...}{Optional arguments for \code{numericInput}.}
}
\description{
Constructs a numeric input with helper.
Same as \code{helper(numericInput(outputId, ...))}
}
\details{
This function uses \code{\link[shinyhelper]{helper}}.
}
