% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{plotOutput_h}
\alias{plotOutput_h}
\title{Plot output with helper}
\usage{
plotOutput_h(outputId, helpfiles = NULL, ...)
}
\arguments{
\item{outputId}{output variable to read the plot/image from.}

\item{helpfiles}{A character string denoting directory to save empty help files.
If \code{NULL} help files will not be created. Default \code{NULL}.}

\item{...}{Optional arguments for \code{plotOutput}.}
}
\description{
Renders plot with helper within an application.
Same as \code{helper(plotOutput(outputId, ...))}
}
\details{
This function uses \code{\link[shinyhelper]{helper}}.
}
