% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_h.R
\name{plotOutput_h}
\alias{plotOutput_h}
\title{Plot output with helper}
\usage{
plotOutput_h(outputId, helpfiles = "helpfiles", ...)
}
\arguments{
\item{outputId}{output variable to read the plot/image from.}

\item{helpfiles}{A character string denoting directory to save empty help files. Default \code{'helpfiles'}.}

\item{...}{Optional arguments for \code{plotOutput}.}
}
\description{
Renders plot with helper within an application.
Same as \code{helper(plotOutput(outputId, ...))}
}
\details{
This function uses \code{\link[shinyhelper]{helper}}.
}
