% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{sliderInput_h}
\alias{sliderInput_h}
\title{Slider input with helper.}
\usage{
sliderInput_h(inputId, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{...}{Optional arguments for \code{sliderInput}.}
}
\description{
Constructs a slider widget with helper.
Same as \code{helper(sliderInput(outputId, ...))}
}
\details{
This function uses \code{\link[shinyhelper]{helper}}.
}
