% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_h.R
\name{sliderInput_h}
\alias{sliderInput_h}
\title{Slider input with helper.}
\usage{
sliderInput_h(inputId, helpfiles = "helpfiles", ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{helpfiles}{A character string denoting directory to save empty help files. Default \code{'helpfiles'}.}

\item{...}{Optional arguments for \code{sliderInput}.}
}
\description{
Constructs a slider widget with helper.
Same as \code{helper(sliderInput(outputId, ...))}
}
\details{
This function uses \code{\link[shinyhelper]{helper}}.
}
