% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tooltip.R
\name{spark_tooltip}
\alias{spark_tooltip}
\title{Generate tooltip}
\usage{
spark_tooltip(hv, plot_obj, plot_type, tt_content, tt_range)
}
\arguments{
\item{hv}{Hoover.}

\item{plot_obj}{Plot to extract data from.}

\item{plot_type}{Type of plot. Accepts either \code{'geom_point'}, \code{'geom_segment'}
or \code{'geom_col'}. Default \code{'geom_point'}.}

\item{tt_content}{Content}
}
\description{
Generates tooltip .
}
