% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{tableOutput_h}
\alias{tableOutput_h}
\title{Table output with helper.}
\usage{
tableOutput_h(outputId, helpfiles = NULL, ...)
}
\arguments{
\item{outputId}{output variable to read the table from}

\item{...}{Optional arguments for \code{dataTableOutput}.}
}
\description{
Renders table with helper within an application.
Same as \code{helper(dataTableOutput(outputId, ...))}
}
\details{
This function uses \code{\link[shinyhelper]{helper}}.
}
