% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{tabsetPanel_SERVER}
\alias{tabsetPanel_SERVER}
\title{Server for tabset panel.}
\usage{
tabsetPanel_SERVER(
  id,
  plot_out,
  table_out,
  plot_type = "point",
  tt_content = NULL
)
}
\arguments{
\item{id}{Id of tabset panel.}

\item{plot_out}{Reactive. Plot to display in tab.}

\item{table_out}{Reactive. Table to display in tab.}

\item{plot_type}{Type of plot. Accepts either \code{'geom_point'}, \code{'comparison'},
\code{'differential'} or \code{'geom_col'}. Default \code{'geom_point'}.}

\item{tt_content}{Optional}
}
\description{
Creates module server for tabset panel with plot and table.
}
