% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{tabsetPanel_UI}
\alias{tabsetPanel_UI}
\title{UI for tabset panel.}
\usage{
tabsetPanel_UI(id)
}
\arguments{
\item{id}{Id name of tabset panel.}
}
\description{
Creates module ui for tabset panel with plot and table.
}
